FIX issue in dl

+++ dl_inv_prob/dl.py
@@ -103,7 +103,7 @@ class DictionaryLearning(nn.Module):
 
     @property
     def D_(self):
-        """ Returns the current dictionary 
+        """ Returns the current dictionary
         np.array (dim_signal, n_components)"""
         return self.D.detach().to("cpu").numpy()
 
@@ -163,7 +163,7 @@ class DictionaryLearning(nn.Module):
         """
         out = torch.zeros(
             (self.n_matrices, self.dim_x, y.shape[2]),
-            dtype=torch.float,
+            dtype=torch.float64,
             device=self.device
         )
 
@@ -373,11 +373,11 @@ class DictionaryLearning(nn.Module):
         if A is None:
             self.dim_signal = self.dim_y
             self.operator = torch.eye(
-                self.dim_y, device=self.device, dtype=torch.float
+                self.dim_y, device=self.device, dtype=torch.float64
             )[None, :]
         else:
             self.dim_signal = A.shape[2]
-            self.operator = torch.from_numpy(A).float().to(self.device)
+            self.operator = torch.from_numpy(A).double().to(self.device)
 
         self.n_matrices = self.operator.shape[0]
 
@@ -385,7 +385,7 @@ class DictionaryLearning(nn.Module):
         if cov_inv is None:
             self.cov_inv = None
         else:
-            self.cov_inv = torch.from_numpy(cov_inv).float().to(self.device)
+            self.cov_inv = torch.from_numpy(cov_inv).double().to(self.device)
 
         # Dictionary
         if self.init_D is None:
@@ -398,10 +398,10 @@ class DictionaryLearning(nn.Module):
             if A is None:
                 dico = Y[0, :, choice]
             self.D = nn.Parameter(
-                torch.tensor(dico, device=self.device, dtype=torch.float)
+                torch.tensor(dico, device=self.device, dtype=torch.float64)
             )
         else:
-            dico_tensor = torch.from_numpy(self.init_D).float().to(self.device)
+            dico_tensor = torch.from_numpy(self.init_D).double().to(self.device)
             self.D = nn.Parameter(dico_tensor)
 
         # Scaling and computing lipschitz
@@ -409,7 +409,7 @@ class DictionaryLearning(nn.Module):
         self.compute_lipschitz()
 
         # Data
-        self.Y_tensor = torch.from_numpy(Y).float().to(self.device)
+        self.Y_tensor = torch.from_numpy(Y).double().to(self.device)
 
         # Training
         loss = self.training_process()
